+++ src/AvroConvert/AvroObjectServices/Write/Resolvers/Record.cs
@@ -20,6 +20,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Linq.Expressions;
+using System.Reflection;
 using SolTechnology.Avro.AvroObjectServices.BuildSchema;
 using SolTechnology.Avro.Features.Serialize;
 using RecordSchema = SolTechnology.Avro.AvroObjectServices.Schema.RecordSchema;
@@ -78,6 +79,13 @@ namespace SolTechnology.Avro.AvroObjectServices.Write.Resolvers
                 if (value == null)
                 {
                     value = type.GetField(name)?.GetValue(recordObj);
+                }
+                if (value == null)
+                {
+                    var privateFields = type.GetFields(BindingFlags.Instance | BindingFlags.NonPublic);
+                    var targetField = privateFields.FirstOrDefault(f => f.Name.Contains(name));
+
+                    value = targetField?.GetValue(recordObj);
                 }
 
                 writer.WriteField(value, encoder);